Fix styling issues
