+++ package.json
@@ -22,6 +22,7 @@
     "@fortawesome/fontawesome-svg-core": "^1.2.18",
     "@fortawesome/free-solid-svg-icons": "^5.8.2",
     "@fortawesome/react-fontawesome": "^0.1.4",
+    "@material-ui/core": "^4.0.0-rc.0",
     "babel-plugin-styled-components": "1.10.0",
     "bootstrap": "^4.3.1",
     "chart.js": "^2.8.0",
