+++ public/index.html
@@ -10,13 +10,6 @@
 
     <link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
 
-    <link
-      rel="stylesheet"
-      href="https://maxcdn.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css"
-      integrity="sha384-ggOyR0iXCbMQv3Xipma34MD+dH/1fQ784/j6cY/iJTQUOhcWr7x9JvoRxT2MZw1T"
-      crossorigin="anonymous"
-      />
-
     <title>IDYEAHS</title>
   </head>
   <body>
