+++ src/app/FeedbackForm.js
@@ -1,37 +1,109 @@
 import React from 'react';
 import styled from 'styled-components';
 
-const StyledForm = styled.form`
+const Grid = styled.div`
   display: grid;
+  justify-items: stretch;
+  align-items: stretch;
   padding: 10px;
   overflow: scroll;
 `;
 
+const StyledForm = styled.form`
+  display: grid;
+  grid-gap: 10px;
+`;
+
 const StyledQuestionContainer = styled.div`
   display: grid;
-  grid-template-rows: 1fr 1fr;
+  grid-template-rows: auto auto auto;
   padding: 10px;
-  margin-bottom: 20px;
-  line-height: 1.4rem;
   color: rebeccapurple;
   background: #efefef;
 `;
 
 const StyledQuestion = styled.div`
+  display: grid;
+  align-items: stretch;
   color: rebeccapurple;
   font-size: 1rem;
-  padding-bottom: 10px;
+  line-height: 1.4rem;
+  padding-bottom: 5px;
 `;
 
-const StyledResponse = styled.div`
-  display: flex;
-  height: 35px;
-  align-content: center;
-  color: rebeccapurple;
-`;
-
-const StyledSubmitButton = styled.button`
+const StyledRadioButtons = styled.div`
   display: grid;
+  grid-template-columns: repeat(5, 1fr);
+  vertical-align: middle;
+  justify-content: center;
+  color: rebeccapurple;
+
+  input {
+    display: none;
+  }
+
+  label {
+    background-color: #efefef;
+    color: rgba(0, 0, 0, 0.6);
+    font-weight: 400;
+    font-size: 1rem;
+    line-height: 1.5rem;
+    text-align: center;
+    padding: 8px 16px;
+    margin-right: 0px;
+    border-top: 1px solid rebeccapurple;
+    border-bottom: 1px solid rebeccapurple;
+    border-right: 1px solid rebeccapurple;
+    transition: all 0.2s ease-in-out;
+  }
+
+  label:hover {
+    cursor: pointer;
+  }
+
+  input:checked + label {
+    background-color: rebeccapurple;
+    color: white;
+  }
+
+  label:first-of-type {
+    border-radius: 4px 0 0 4px;
+    border-left: 1px solid rebeccapurple;
+  }
+
+  label:last-of-type {
+    border-radius: 0 4px 4px 0;
+    border-right: 1px solid rebeccapurple;
+  }
+`;
+
+const StyledLegend = styled.div`
+  display: grid;
+  grid-template-columns: 1fr 1fr 1fr;
+  align-items: center;
+  font-size: 0.7rem;
+  margin: 0;
+  padding: 5px 0 5px 0;
+  color: #a2a2a2;
+
+  div:first-child {
+    display: grid;
+    justify-items: start;
+  }
+
+  div:nth-child(2) {
+    display: grid;
+    justify-items: center;
+  }
+
+  div:last-child {
+    display: grid;
+    justify-items: end;
+  }
+`;
+
+const StyledButton = styled.button`
+  display: flex;
   margin-bottom: 5px;
   padding: 5px;
   background: #dacc3e;
@@ -45,63 +117,73 @@ export default function FeedbackForm({ questions, handleSubmit, history }) {
   function handleFormSubmit(event, questions, history) {
     event.preventDefault();
     const form = event.target;
-    const answers = questions.map(question => form[`q-${question.id}`].value);
+    const answers = questions.map(
+      question => form[`q-${question.id + 1}`].value
+    );
     console.log('Answers:', answers);
     handleSubmit(answers, history);
   }
 
   return (
-    <StyledForm
-      id="addFeedback"
-      onSubmit={event => handleFormSubmit(event, questions, history)}
-    >
-      {questions.map(question => (
-        <StyledQuestionContainer key={question.id}>
-          <StyledQuestion>
-            <strong>Q{question.id + 1}.</strong> {question.question_text_en}
-          </StyledQuestion>
-          <StyledResponse>
-            <input
-              type="radio"
-              name={`q-${question.id}`}
-              id="radio-1"
-              value="1"
-            />
-            <label htmlFor="radio-1">strongly disagree</label>
-            <input
-              type="radio"
-              name={`q-${question.id}`}
-              id="radio-2"
-              value="2"
-              defaultChecked={true}
-            />
-            <label htmlFor="radio-2">disagree</label>
-            <input
-              type="radio"
-              name={`q-${question.id}`}
-              id="radio-3"
-              value="3"
-            />
-            <label htmlFor="radio-3">neutral</label>
-            <input
-              type="radio"
-              name={`q-${question.id}`}
-              id="radio-4"
-              value="4"
-            />
-            <label htmlFor="radio-4">agree</label>
-            <input
-              type="radio"
-              name={`q-${question.id}`}
-              id="radio-5"
-              value="5"
-            />
-            <label htmlFor="radio-5">strongly agree</label>
-          </StyledResponse>
-        </StyledQuestionContainer>
-      ))}
+    <Grid>
+      <StyledForm
+        id="addFeedback"
+        onSubmit={event => handleFormSubmit(event, questions, history)}
+      >
+        {questions.map(question => (
+          <StyledQuestionContainer key={question.id}>
+            <StyledQuestion>
+              Q{question.id + 1}. {question.question_text_en}
+            </StyledQuestion>
 
-      <StyledSubmitButton>Add Feedback</StyledSubmitButton>
-    </StyledForm>
+            <StyledRadioButtons>
+              <input
+                type="radio"
+                name={`q-${question.id + 1}`}
+                id={`radio-${question.id + 'a'}`}
+                value="1"
+              />
+              <label htmlFor={`radio-${question.id + 'a'}`}>1</label>
+              <input
+                type="radio"
+                name={`q-${question.id + 1}`}
+                id={`radio-${question.id + 'b'}`}
+                value="2"
+              />
+              <label htmlFor={`radio-${question.id + 'b'}`}>2</label>
+              <input
+                type="radio"
+                name={`q-${question.id + 1}`}
+                id={`radio-${question.id + 'c'}`}
+                value="3"
+                defaultChecked={true}
+              />
+              <label htmlFor={`radio-${question.id + 'c'}`}>3</label>
+              <input
+                type="radio"
+                name={`q-${question.id + 1}`}
+                id={`radio-${question.id + 'd'}`}
+                value="4"
+              />
+              <label htmlFor={`radio-${question.id + 'd'}`}>4</label>
+              <input
+                type="radio"
+                name={`q-${question.id + 1}`}
+                id={`radio-${question.id + 'e'}`}
+                value="5"
+              />
+              <label htmlFor={`radio-${question.id + 'e'}`}>5</label>
+            </StyledRadioButtons>
+            <StyledLegend>
+              <div>strongly disagree</div>
+              <div>neutral</div>
+              <div>strongly agree</div>
+            </StyledLegend>
+          </StyledQuestionContainer>
+        ))}
+
+        <StyledButton>Add Feedback</StyledButton>
+      </StyledForm>
+    </Grid>
   );
 }
