+++ src/app/Header.js
@@ -4,13 +4,14 @@ import { NavLink } from 'react-router-dom';
 
 const Heading = styled.header`
   display: grid;
+  justify-content: center;
+  align-items: center;
   height: 80px;
   background: rgba(127, 183, 190, 0.8);
 `;
 
 const Title = styled.h1`
   display: grid;
-  align-self: center;
 
   color: white;
   text-align: center;
